fix: re-introduce paramsSerializer (#408)
We reintroduce the `paramsSerializer` (again) after removing it
[here](https://github.com/contentful/contentful-sdk-core/pull/370).

This is to unblock [the axios bump in
contentful-merge](https://github.com/contentful/contentful-merge/pull/462)
or the skd-core bump [as
such](https://github.com/contentful/contentful-merge/pull/552). We saw
that tests were failing because the query set together
[here](https://github.com/contentful/contentful-merge/blob/main/src/engine/create-changeset/tasks/create-fetch-partial-entities-task.ts#L43)
can not properly be handled by the API anymore and instead runs into

`error: InvalidQuery: The query you sent was invalid. Probably a filter
or ordering specification is not applicable to the type of a field.`

Tested locally on `contentful.js` and `contentful-merge`